pdf 2 jpg/pdf-to-jpg-exe/requirements.txt
Pillow==9.0.1
pdf2image==1.16.0
tkinter==0.1.0